add command menu to its own component
